Move ResourceSegmentsConverter from sharding-distsql-handler module to distsql-statement module
