+++ distsql/statement/src/test/java/org/apache/shardingsphere/distsql/parser/segment/converter/ResourceSegmentsConverterTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.distsql.handler.converter;
+package org.apache.shardingsphere.distsql.parser.segment.converter;
 
 import org.apache.shardingsphere.distsql.parser.segment.DataSourceSegment;
 import org.apache.shardingsphere.distsql.parser.segment.HostnameAndPortBasedDataSourceSegment;
